Force apiary.apib rebuild in Travis checks
Make relies on file modification timestamps, and when people work
concurrently in the project, it's not reliable.  For instance,
apiary.apib may be lacking some changes from a source file, and still be
newer than that source file because it has been updated in another pull
request.

For this reason, always rebuild apiary.apib in Travis before doing other
checks.